+++ apps/ai-dial-admin/src/components/Grid/CellRenderers/TurnIdCellRenderer.tsx
@@ -2,7 +2,8 @@ import { ICellRendererParams } from 'ag-grid-community';
 
 import { DialEllipsisTooltip } from '@epam/ai-dial-ui-kit';
 
-import { GridRowType, GroupedGridRow } from '@/src/models/evaluation/test-case-grouping';
+import { GroupedGridRow } from '@/src/models/evaluation/test-case-grouping';
+import { GridRowType } from '@/src/types/grid-row-type';
 
 const TurnIdCellRenderer = ({ data }: ICellRendererParams<GroupedGridRow>) => {
   if (data?.rowType === GridRowType.TURN && !data.isFlattened) return null;
