+++ apps/ai-dial-admin/src/types/grid-cell.ts
