+++ apps/ai-dial-admin/src/types/grid-row-type.ts
@@ -1,0 +1,5 @@
+export enum GridRowType {
+  GROUP = 'GROUP',
+  TURN = 'TURN',
+  SINGLE = 'SINGLE',
+}
