+++ apps/ai-dial-admin/src/types/test-case-row.ts
@@ -1,0 +1,1 @@
+export type TestCaseRow = Record<string, unknown>;
